requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
flask-mail  # For sending emails
pyjwt       # For expiring tokens (optional)
python-dotenv  # For email service credentials
bcrypt      # For password hashing
flask_sqlalchemy
flask_login
flask_bcrypt
requests
google_auth_oauthlib
git-filter-repo
flask-wtf
pip

